Secure_Terminal/requirements.txt
flask
flask-socketio
pycryptodome
